PreviewPageFrame checked state updates
Send the new checked event. Updated to the new checked api.

This widget is using getSelected() as the checked state. We can't send
the event in setSelected() because setSelected in called in
onBindViewHolder and the widget is not attached yet. So we send the
new event when the widget is clicked.

Test: tested on the print preview and checked the new event is sent
when clicked.
Flag: android.view.accessibility.tri_state_checked
Bug: 380327448

Change-Id: I1591425bf7d35ef361cae94ee394cb26b20014ff