Allow search kcm fallbacks
Fixes a bug in where kcm fallbacks with search based modifiers could
not activate since PhoneWindowManager would consume all search events.
This prevented the InputDispatcher from generating the fallback.

This change modifies the behavior so that when a search-based
fallback is detected, the key event will be directly sent to
InputDispatcher to handle the fallback. This will NOT send the
original event to applications.

Bug: 384113980
Test: atest PhoneWindowManagerTest
Test: atest InputTests:InputManagerServiceTests

Flag: com.android.hardware.input.fix_search_modifier_fallbacks

Change-Id: Iacecf945ea6be9841b6009d5f30bcf83f68309ef