[flexiglass] Keep the snoozable HUN placeholder scrollable
Previously we were only adding the scrollable modifier to the
HeadsUpNotificationSpace placeholder when there was a pinned HUN
displayed, or being animated away.
When we are dragging down on a HUN to open the shade, we unpin it, to
let it reach its final form in the shade. This unpinning removed the
scrollable modifier, which broke dispatching the touch events from
NestedDraggable to the DraggableHandler. This was leaving the HUN and
the shade in a broken state.

Keeping the snoozable placeholder scrollable allows us to keep dispatching
the motion events continously, and open the shade.

Fixes: 394800311
Test: atest PlatformScenarioTests:android.platform.test.scenario.sysui.notification.ExpandShadeFromHun#testExpandShadeFromHun
Flag: com.android.systemui.scene_container
Change-Id: I891adfad8da40e21852e125e59024c4302fda026