Provide NotificationStackModule also in aosp
NotificationStackModule was provided only in the NotificationsGoogleModule, but also in aosp we're using NotificationStackScrollLayout, so we should have this module there as well.

I realized of the issue while trying to turn on shade_window_goes_away flag: in aosp the NotificationStackRebindingHider was not being provided, making sysui crash with the flag on.

Bug: 362719719
Test: run some tests on aosp emulator sucessfully with this and the flag on
Flag: com.android.systemui.shade_window_goes_around
Change-Id: I06527b215a67de4f507ddbc48131b687c50d9df9