Switch to gesture nav in InputMethodServiceTest
Previously the tests related to (long) clicking on the IME back button
and IME Switcher button were skipped if the device was not already in
gesture navigation. It turns out the test is currently configured to
only run on targets which are always in three button navigation, so the
test is never actually ran.

This uses the CTS utility to enable switching to gesture navigation mode
if needed. To avoid re-creating the Activity when the mode changes, this
marks that the TestActivity handles assetsPaths config changes.

To avoid the countDownLatch being triggered by the IME receiving the
navigation mode config change, this explicitly waits for it when
necessary.

Flag: EXEMPT testfix
Bug: 394548201
Test: atest InputMethodServiceTest#testBackButtonClick
  InputMethodServiceTest#testBackButtonLongClick
  InputMethodServiceTest#testImeSwitchButtonClick
  InputMethodServiceTest#testImeSwitchButtonLongClick
Change-Id: Ib8ab74918490ffa6a5d382c3dc782a3e3039970c