Add a flag for opt-out for camera compat for freeform.
This is a flag for the behavior change (check otp out property), not for exposing the API - this will be guarded with a different flag that will require API review and respect API timelines. The flag will be exposed via Jetpack for an earlier launch.

Flag: com.android.window.flags.enable_camera_compat_for_desktop_windowing_opt_out
Test: m
Bug: 328616176
Change-Id: I2306614eef0f14a7bcd585f4b5237cdc6eb5a62a