[PiP2] Implement expand PiP to Split CUJ
Whenever PiP enters from a multi-activity task,
the last parent before PiP can enter split-screen.
So we detect whether that's happened and expand into the split
instead of fullscreen, depending on the context.

This is also the first CL that is cleaning up PipTransition by
separating it into smaller handlers. Here we are adding a new
PipExpandHandler that PipTransition delegates the transition resolution
and animation to if needed.

Also added a new PipExpandHandlerTest in the effort to make PiP
transitions more covered by unit tests.

Bug: 381330052
Flag: com.android.wm.shell.enable_pip2
Test: enter multi-activity PiP from split, then expand from PiP
Test: atest WMShellUnitTests:PipSchedulerTest
Test: atest WMShellUnitTests:PipExpandHandlerTest
Change-Id: I660176dc1a247613aafdf1fa06db3dfec10ad597