Init workload target in java FrameInfo
FrameInfo's workload target was never initialized on the Java
side, and it resulted in occasions where the workload target
value was 0. FrameInfo in DrawFrameTask populates itself from
the FrameInfo.java, and this info is propagated into CanvasContext.
This change ensures that the workload target also has a value
going into JankTracker.

Fixes: b/395917245
Test: presubmit, wear metric tests
Flag: EXEMPT, bug fix
Change-Id: I6540ee66d1ad8b0a22da516120cb2b9135c70398