Use DesktopExperienceFlag for ShadeWindowGoesAround
This allows to enable it with all other flags from the dev setting.

Note that the shade will not move even if the flag is enabled unless the
user selects a "shade display policy" from dev settings or with the adb
command.

Bug: 362719719
Test: The flag gets enabled with the dev option
Flag: com.android.systemui.shade_window_goes_around
Change-Id: Ie4716026588b1bd164ef8f04673c009a9dceb4d1