Workaround to enable sysui lint checks in ASfP
System UI's custom lint checks will not run in ASfP unless the jar path
to the checks is specified in the lint.xml of the module.

This is a temporary workaround to get lint checks to work in ASfP when
using the module path "frameworks/base/packages/SystemUI".

Bug: 397519800
Bug: 398194411
Test: Create ASfP project, look for custom inline lint inspections
Test: m SystemUI-core-lint and view report
Flag: EXEMPT lint only
Change-Id: I92e2febd39a96cb20ca17296f2ae093217b6dbb2