Make SysUI state instances display aware
With this cl each instance of SysUIState will be associated to a
specific display id.
The default instance provided in the dagger graph is for the default
display. In a follow up, support for multiple instance will be added,
and propagated to launcher.

In this cl DisplayTracker is removed, as the default display is always
Display.DEFAULT_DISPLAY (and will always be that). The old
"onSystemUiStateChanged" is only propagated for the default display also
 after this change.

Bug: 362719719
Bug: 398011576
Test: SysUiStateTest
Flag: com.android.systemui.shade_window_goes_around
Change-Id: Ie9b3b9f57848a92e25db389ee4cb18ae308d679e