[3/N] Report home visibility update on cancelling DragToDesktop transition
When the DragToDesktop START transition gets interrupted (from an unknown
incoming transition) before we've requested an END/CANCEL transition, we
finish the START transition and let it move the user to the Home screen.
In this case we never request an END/CANCEL transition, and before this
CL we would thus never update Home visibility. With this CL we update the
Home visibility when START ends (onTransitionFinished) if we haven't
received an END/CANCEL transition by then.

Bug: 397135730
Test: HomeTransitionObserverTest
Flag: com.android.window.flags.enable_drag_to_desktop_incoming_transitions_bugfix
Change-Id: I7b7c0f40f274210387253a6f02d1eb471b0332b1