Hook up drag-to-fullscreen from the bubble bar expanded view
This is behind the fullscreen flag and uses the new drop
target manager to show the drop zones for moving left/right
and to fullscreen.

This is probably not the ideal way to do things, but I wanted
to get this behavior in the build for next week!

Flag: com.android.wm.shell.enable_bubble_to_fullscreen
Test: atest DropTargetManagerTest
Test: manual - drag the bubble expanded view around and
               observe the drop targets
Test: manual - drag the bubble expanded view to the opposite
               side without releasing, drag to fullscreen and
               release, get a new bubble => make sure it's on
               the correct side
Bug: 388857737
Change-Id: Id07e4bbfc79766124f7e162d020181e451baece8