[34/N] Desks: Implement task-limit policy
Adds the missing task-limit and minimization policies needed when a task
is being added to a desk that weren't functional with multiple desks
enabled.

Instead of relying on the task to minimize being left behind Home,
DesksOrganizer#minimizeTask is used to send the task to the minimization
container of its desk.

Flag: com.android.window.flags.enable_multiple_desktops_backend
Bug: 391485148
Test: Move a task to a desk that's at the limit, verify it minimizes the
one at the back.

Change-Id: Id4773618dc62a667b2deb9268f7d6d9c11287d20