Fix flicker in taskbar visibility when shade moves display
SysUI was telling to launcher the shade was visible in the old display when the expansion on an external display happened.
This was triggering the taskbar to become invisible for a few frames, and then return visible on the default display.

By using the shade repository pending id we can immediately propagate the SysUIState for the correct display

Bug: 362719719
Bug: 398011576
Test: NotificationPanelViewControllerTest
Flag: com.android.systemui.shade_window_goes_around
Change-Id: I7eab3284acf9b295d32fd2c0802974e1cc562611