[Notif redesign] Allow collapsed custom views to take up to 48dp
Normally collapsed notification content has a height of 40dp, with a
margin of 16dp above and below (so the standard height for collapsed
notifications is 72dp in the redesign). However, custom views have
previously been allowed to have a height of up to 48dp when collapsed,
so the new design was causing them to get cut off.

In order to fix this issue without directly making custom view
notifications taller, we can reduce the margin to 12dp (while applying
the additional 4dp to the child views, for the non-custom-view case).

Test: manual for custom views, plus screenshot tests should be unchanged
Flag: android.app.notifications_redesign_templates
Bug: 378660052
Fix: 386397001
Change-Id: I89a2c82c8f4b022d2bb8a50483cdf6c5c4aac2a3