Add deferTransitionProgress flag to STL (1/2)
This CL adds a new deferTransitionProgress flag to STL to reduce
perceivable jank by waiting for the first expensive composition frame of
a transition to be done before changing the progress of the transition.
This works with gesture-based transitions and one-off transitions.

See b/400688335#comment2 for details.

Bug: 400688335
Test: atest NestedDraggableTest
Flag: EXEMPT new behavior disabled by default
Change-Id: I839529e2bc1d792ae8a37d5ddf3e7c61a12fe3cb