Create task drag zones for foldables
Foldables or small tablets do not have desktop available.
If the screen where drag is taking place is a small tablet, use a
different set of drag zones compared to large tablets.

Initialize the list of regions for small tablets when creating the
indicator. This ensures that we don't need to recalculate regions on
each drag event.

Put calculated regions (rect area to indicator type) to a sorted list.
Once a motion event is received match the coordinates to the list of
regions. The first rect that contains this event will determine the
indicator type.

Bug: 397697080
Test: atest DesktopModeVisualIndicatorTest
Test: manual, on fold drag a fullscreen task to split or bubble
Test: manual, on desktop, drag a fullscreen task to desktop or split or
  bubble, drag a desktop task to fullscreen
Flag: com.android.wm.shell.enable_bubble_to_fullscreen
Change-Id: I30d69530f196abb375d250df5e72b18b1e9e7dd4