a11y: Fix autoclick typePanel buttons only work when clicking on edge
Root cause: The click listener is set on parent LinearLayout, When
ImageButton gets the click, it does not propagate to parent, instead it
consumes the click event, preventing the event from reaching the parent
LinearLayout.

Fix: Set android:clickable="false" on the ImageButton in XML to allow
the event to propagate to the parent.

Video:
 - Before: http://shortn/_LMnTIsoIVY
 - After:  http://shortn/_S8Wx9UC1z9

Bug: b/395947503
Test: AutoclickTypePanelTest
Flag: com.android.server.accessibility.enable_autoclick_indicator
Change-Id: Id389af6539f92fd066e5e6f7141aeb7a39a47b5b