[Metrics] Use the correct cutoff param
The DELTA_EXCEPTION_CUT is meant for exceptions, and was accidentally
used instead of DELTA_RESPONSES_CUT. The CL was locally ready, but it
was packaged along with another change that didn't have to be made, so
it was overlooked to patch. This completes the missing relevant bit that
was left out when deciding another bit didn't need to be changed.

Bug: 364444783
Test: Build, stats_d already tested shows this fixes longer string.
Change-Id: Id246ace7c4372342cb36d6956edf68a5497eb643
Flag: android.credentials.flags.fix_metric_duplication_emits