Ensure dream logic is not suspended on handler.
It is possible for work to be queued on a handler and then
suspended in doze mode. This can cause dream operations to
be noticeably deferred. In order to address this, a wake
lock is now acquired in this scenario and wraps these
calls on the handler. Note that in the normal, non-doze
dream case, the wakelock is not necessary.

Bug: 404420534
Flag: EXEMPT bugfix
Test: atest DreamServiceTest
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:d958722427762581b9a39b3518d21334e063d99b)
Merged-In: Ib7802e951ea793ca4484ef19972a9f2aed4bfd81
Change-Id: Ib7802e951ea793ca4484ef19972a9f2aed4bfd81